+++ clj/CHANGELOG.md
@@ -16,8 +16,9 @@ All notable changes to this project will be documented in this file. This change
 
 ### Changed
 
-- `rollup` must now be applied to an inventory, to facilitate use across different standard queries.
+- `rollup` must now be applied to an inventory, to facilitate use across different standard queries
 - journal now includes posting flags, if any
+- only `*beans*` is used by the top-level queries, with `*directives*` et al being ignored
 
 [commit log]: https://github.com/tesujimath/limabean/compare/0.4.1...HEAD